moved some code into different files; Still having some problems with recovered image discoloration
