+++ bch_in_Cpp/CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.10)
 project(BCH_code)
 
-set(CMAKE_BUILD_TYPE Release)
+set(CMAKE_BUILD_TYPE Debug)
 
 set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
@@ -10,6 +10,8 @@ set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ..)
 
 set(BCH6351_SOURCE
      bch6351.cpp
+     bch_utils.cpp
+     bch_math.cpp
      bch_logger.hpp)
 
 # set(BCH6345_SOURCE
